chnage var position and clean code

+++ .github/action/utils/sonar-scanner.sh
@@ -4,15 +4,15 @@
 # *********************************************************************
 #
 # Setting variable
-SONAR_URL       = %1
-SONAR_USER      = %2
-SONAR_PASS      = %3
-SONAR_LANGUAGE  = %4
-SONAR_CLI       = %5
-REF             = %6
-GROUPID         = %7
-ARTIFACTID      = %8
-VERSION         = %9
+SONAR_URL       = ${{ env.SONAR_URL }}
+SONAR_USER      = ${{ env.SONAR_USER }}
+SONAR_PASS      = ${{ env.SONAR_SECRET }} 
+SONAR_LANGUAGE  = ${{ env.LANGUAGE }}
+SONAR_CLI       = ${{ env.SONAR_CLI }}
+REF             = ${{ github.ref }}
+GROUPID         = ${{ github.GROUP }}
+ARTIFACTID      = ${{ github.ARTIFACT }}
+VERSION         = ${{ github.VERSION }} 
 
 # if url isn't empty then allow sonar for scanner code
 if [ ${SONAR_URL} != "" && ${ startsWith(${ REF }, 'refs/heads/main') } == true ] then  
@@ -22,17 +22,25 @@ if [ ${SONAR_URL} != "" && ${ startsWith(${ REF }, 'refs/heads/main') } == true 
   echo "***************************************************"
   
   # install client version
-  apt-get update
-  apt-get install unzip wget nodejs
-  mkdir /downloads/sonarqube -p
-  cd /downloads/sonarqube
-  wget https://binaries.sonarsource.com/Distribution/sonar-scanner-cli/sonar-scanner-cli-${SONAR_CLI}-linux.zip
-  unzip sonar-scanner-cli-${SONAR_CLI}-linux.zip
-  mv sonar-scanner-${SONAR_CLI}-linux /opt/sonar-scanner
-  rm -r -f /downloads/sonarqube
-  export PATH="$PATH:/opt/sonar-scanner/bin"
-  # env | grep PATH
   sonar-scanner -v 
+  if echo $? = 128 ; then
+      echo "The program 'sonar-scanner' is currently not installed."
+      echo "install program 'sonar-scanner ${SONAR_CLI}'..."
+      apt-get update
+      apt-get install unzip wget nodejs
+      mkdir /downloads/sonarqube -p
+      cd /downloads/sonarqube
+      wget https://binaries.sonarsource.com/Distribution/sonar-scanner-cli/sonar-scanner-cli-${SONAR_CLI}-linux.zip
+      unzip sonar-scanner-cli-${SONAR_CLI}-linux.zip
+      mv sonar-scanner-${SONAR_CLI}-linux /opt/sonar-scanner
+      rm -r -f /downloads/sonarqube
+      export PATH="$PATH:/opt/sonar-scanner/bin"
+      # env | grep PATH
+      sonar-scanner -v 
+  else
+      echo "Continuing with sonar-scanner."
+  fi
+ 
 
   if [${SONAR_LANGUAGE}=="java"] then
